requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
appdirs==1.4.4
argcomplete==3.0.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
boto==2.49.0
cachetools==5.3.0
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
crcmod==1.7
cryptography==39.0.2
cycler==0.11.0
dgl==0.4.3
dgl-cu102==0.6.1
docker-pycreds==0.4.0
fasteners==0.18
fonttools==4.38.0
frozenlist==1.3.3
gcs-oauth2-boto-plugin==3.0
gitdb==4.0.10
GitPython==3.1.31
google-apitools==0.5.32
google-auth==2.16.2
google-reauth==0.1.1
gsutil==5.21
gym==0.26.2
gym-notices==0.0.8
httplib2==0.20.4
idna==3.4
importlib-metadata==5.2.0
joblib==1.2.0
kiwisolver==1.4.4
learn2learn==0.1.7
matplotlib==3.5.3
monotonic==1.6
multidict==6.0.4
networkx==2.6.3
numpy==1.21.6
oauth2client==4.1.3
packaging==23.0
pandas==1.3.5
pathtools==0.1.2
Pillow==9.4.0
protobuf==4.22.1
psutil==5.9.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.2
pyu2f==0.1.5
PyYAML==6.0
qpth==0.0.15
requests==2.28.2
retry-decorator==1.1.1
rsa==4.7.2
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.17.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
torch==1.9.0+cu102
torchaudio==0.9.0
torchvision==0.10.0+cu102
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
wandb==0.14.0
yarl==1.8.2
zipp==3.15.0
